tach logic PlayerController sang CharacterBase
