+++ Assets/Prefabs/Bullet.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bfb03c391771f1e4c99baa2a61d916df
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
