+++ Assets/Scenes/Ingame.unity
@@ -408,7 +408,6 @@ GameObject:
   - component: {fileID: 650376207}
   - component: {fileID: 650376206}
   - component: {fileID: 650376205}
-  - component: {fileID: 650376204}
   m_Layer: 6
   m_Name: Player
   m_TagString: Untagged
@@ -416,21 +415,6 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &650376204
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 650376203}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a21048edc1d6a5c4590a9ed9165cb299, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  shootPoint: {fileID: 1009661173}
-  bulletPrefab: {fileID: 6510611370532917505, guid: 18e76720ba51e4643bc485af3fa10446, type: 3}
-  _shootTime: 1
 --- !u!114 &650376205
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -451,7 +435,10 @@ MonoBehaviour:
     serializedVersion: 2
     m_Bits: 128
   groundCheck: {fileID: 471805959}
-  player_dir: 1
+  character_dir: 1
+  shootPoint: {fileID: 1009661173}
+  bulletPrefab: {fileID: 6510611370532917505, guid: 18e76720ba51e4643bc485af3fa10446, type: 3}
+  _shootTime: 0.3
 --- !u!61 &650376206
 BoxCollider2D:
   m_ObjectHideFlags: 0
