+++ Assets/Scripts/Controller/Player/PlayerController.cs
@@ -4,25 +4,27 @@ using UnityEngine;
 
 public class PlayerController : CharacterBase
 {
-    
-
     private float inputX;
-    public int player_dir = 1; // Huong cua nhan vat, de tinh toan huong cua vien dan
-    private bool facing_right = true;
 
+    [Header("Shoot")]
+    public Transform shootPoint;
+    public GameObject bulletPrefab;
+    [SerializeField] private float _shootTime;
+    float _shootTimeCount = 0;
+
+    [Header("Bullet available")]
     private Dictionary<Color, int> absorbedBullets = new();
 
-    void Start()
+    protected override void Start()
     {
-        rb = GetComponent<Rigidbody2D>();
-        collid = GetComponent<BoxCollider2D>();
-        
+        base.Start();
     }
 
     void Update()
     {
         InputControl();
         FlipController();
+        TestShoot();
     }
 
     private void InputControl()
@@ -61,13 +63,6 @@ public class PlayerController : CharacterBase
             rb.AddForce(new Vector2(rb.velocity.x, jumpForce), ForceMode2D.Impulse);
     }
 
-    private void Flip()
-    {
-        player_dir *= -1;
-        facing_right = !facing_right;
-        rb.transform.Rotate(0, 180, 0);
-    }
-
     private void FlipController()
     {
         if (inputX == 1 && !facing_right)
@@ -76,48 +71,25 @@ public class PlayerController : CharacterBase
             Flip();
     }
 
-    private void OnCollisionEnter2D(Collision2D collision)
+    private void TestShoot()
     {
-        if (collision.gameObject.CompareTag(GameConstants.oneWayPlatform) && !platformCantFall.Contains(collision.gameObject))
+        if (_shootTimeCount > 0)
         {
-            currentOneWayPlatform = collision.gameObject;
+            _shootTimeCount -= Time.deltaTime;
+            return;
         }
-    }
 
-    private void OnCollisionExit2D(Collision2D collision)
-    {
-        if (collision.gameObject.CompareTag(GameConstants.oneWayPlatform))
+        if (Input.GetKeyDown(KeyCode.C))
         {
-            currentOneWayPlatform = null;
-        }
-    }
+            GameObject g = ObjectPool.Instance.GetObject(bulletPrefab);
+            g.transform.position = shootPoint.position;
 
-    private IEnumerator DisableCollision()
-    {
-        BoxCollider2D platformCollider = currentOneWayPlatform.GetComponent<BoxCollider2D>();
+            float dir = character_dir;
+            g.transform.rotation = Quaternion.Euler(0, (dir == 1 ? 0 : -180), 0);
 
-        Physics2D.IgnoreCollision(collid, platformCollider);
-        yield return new WaitForSeconds(1.0f);
-        Physics2D.IgnoreCollision(collid, platformCollider, false);
-    }
+            g.SetActive(true);
 
-    private bool IsGround()
-    {
-        RaycastHit2D hit = Physics2D.Raycast(groundCheck.position, Vector2.down, 0.05f, groundLayer);
-        return hit.collider != null;
-    }
-
-    private void OnDrawGizmos()
-    {
-        Gizmos.color = Color.yellow;
-
-        RaycastHit2D hit = Physics2D.Raycast(groundCheck.position, Vector2.down, 0.05f, groundLayer);
-
-        Gizmos.DrawRay(groundCheck.position, Vector2.down * 0.05f);
-
-        if (hit.collider != null)
-        {
-            Gizmos.DrawSphere(hit.point, 0.05f);
+            _shootTimeCount = _shootTime;
         }
     }
 
@@ -129,7 +101,7 @@ public class PlayerController : CharacterBase
         }
         else if (currentColor != bulletColor)
         {
-            HandleGameOver();
+            HandleGameOver(); // Ham nay xu ly trong GameManager, nhung tam thoi de o day
             return;
         }
