--- Assets/Scripts/Controller/Player/PlayerShooting.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a21048edc1d6a5c4590a9ed9165cb299
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
